Add shadow support
There are a few tricks to creating good looking shadows:

* Stacking shadows on top of each other to create a more natural look
* Color-matching: Make the shadow color a desaturated version of the background color
* Using consistent 'shadow elevation'

Fulfilling all of these criterias turned out to be more tricky than anticipated. Hard coding shadows in the theme file would be easy, but creating consistent shadows on demand is hard.

The solution? The utility function createShadow(size, color).

createShadow() fulfills all of the critera above. Since it can be used both when defining the theme and on demand in Stitches, it makes it easy to create consistent shadows.

I had to create quite a few utility functions and types to make all of
this work:
* transparentize(): Changes the opacity of theme colors
* desaturate(): Changes the saturation of theme colors. While the theme colors are great, they must be desaturated before they can be used as shadows
* hasKey(): Checks if an object contains a given key. Needed to avoid Typescript errors
* type LiteralUnion: In types like `'dog' | 'cat' | string`, the autocomplete for 'dog' and 'cat' would break. By using LiteralUnion we can keep the great automplete.

Lastly: 'Polished' is added as a dependency, as it powers the color utility functions.